Apache 2.0 license
This project is licensed under the Apache License 2.0.

I have chosen Apache 2.0 because in a previous job I was deprived of proper authorship for my work. 
The experience of having my contributions overlooked was extremely unfair. 
I hope this license ensures that all contributions are properly attributed, 
and that credit is given where it is due.